requirements.txt
# FastAPI и связанные зависимости
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# OpenAI Agents
openai-agents==0.2.6

# Переменные окружения
python-dotenv==1.0.0

# Для разработки
pytest>=8.0.0
pytest-asyncio>=0.24.0
